Split command module up for better readability
